refactor: separate search artist and concert logic
