+++ src/api/index.ts
@@ -2,3 +2,4 @@ export * from './refundAccountApi';
 export * from './rentalApi';
 export * from './surveyApi';
 export * from './rentalFormApi';
+export * from './searchApi';
