+++ src/queries/search/useGetSearchArtist.ts
