+++ src/types/index.ts
@@ -5,3 +5,4 @@ export * from './survey';
 export * from './user';
 export * from './api';
 export * from './filter';
+export * from './search';
